Check where we should be broadcasting the mode to the channel Fixes #10 closed.
sendChannelModeReply was sending over a command instead of the code the client expected.

Also improved the info fetching so additional spaces don't get added to the mode  string.

Included adding booleeans inside of the switch case  in parseAndApplyChannelModes.

Channel limits are now parssed into an unsigned long as IRC does not support negative values in parameters. This change makes it so negative values are treated as errors and mode adjustment is ignored.